+++ modules/terrain/src/lib/parse-terrain.js
@@ -1,4 +1,5 @@
 import Martini from '@mapbox/martini';
+import {ImageLoader} from '@loaders.gl/images';
 
 function getTerrain(imageData, tileSize, elevationDecoder) {
   const {rScaler, bScaler, gScaler, offset} = elevationDecoder;
@@ -89,10 +90,9 @@ function getMartiniTileMesh(terrainImage, terrainOptions) {
 }
 
 export default async function loadTerrain(arrayBuffer, options, context) {
-  options.image = {
-    type: 'data'
-  };
-  const image = await context.parse(arrayBuffer, options, options.baseUri);
+  options.image = options.image || {};
+  options.image.type = 'data';
+  const image = await context.parse(arrayBuffer, ImageLoader, options, options.baseUri);
   // Extend function to support additional mesh generation options (square grid or delatin)
   return getMartiniTileMesh(image, options.terrain);
 }